requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
event-bus==1.0.2